requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
Pillow>=11.0.0
